+++ libs/openai/src/lib/base.ts
@@ -1,5 +1,7 @@
 import axios, { AxiosRequestConfig } from 'axios'
-import { ListEngine, AnswerResponse, AnswerRequest, ListFile, File } from '../models'
+import { ListEngine, AnswerResponse, AnswerRequest, ListFile, File, FileRequest } from '../models'
+import * as FormData from 'form-data'
+import * as fs from 'fs'
 
 export class OpenAI {
   protected apiKey: string
@@ -9,7 +11,7 @@ export class OpenAI {
     this.apiKey = apiKey
   }
 
-  private async request<T>(url: string, method: 'GET' | 'POST', data?: AnswerRequest): Promise<T> {
+  private async request<T>(url: string, method: 'GET' | 'POST', data?: AnswerRequest | FileRequest): Promise<T> {
     try {
       const options: AxiosRequestConfig = {
         method,
@@ -19,6 +21,21 @@ export class OpenAI {
           'Content-Type': 'application/json',
         },
         data: data || '',
+      }
+
+      // TODO - better type checking for data
+      // Upload file
+      if (data.file != null && data['purpose'] != null) {
+        const formData = new FormData()
+        formData.append('purpose', data['purpose'])
+        formData.append('file', fs.createReadStream(data.file))
+
+        options.headers = {
+          ...options.headers,
+          ...formData.getHeaders(),
+        }
+
+        options.data = formData
       }
 
       const response = await axios(options)
@@ -37,11 +54,17 @@ export class OpenAI {
     return this.request<AnswerResponse>(`${this.baseUrl}/answers`, 'POST', data)
   }
 
+  // FILES
+
   listFiles(): Promise<ListFile> {
     return this.request<ListFile>(`${this.baseUrl}/files`, 'GET')
   }
 
   retrieveFile(fileId: string): Promise<File> {
     return this.request<File>(`${this.baseUrl}/files/${fileId}`, 'GET')
+  }
+
+  uploadFile(data: FileRequest): Promise<File> {
+    return this.request<File>(`${this.baseUrl}/files`, 'POST', data)
   }
 }